Updated to Vapor 4.
Largely followed the recipe for upgrading shown here: https://docs.vapor.codes/4.0/upgrading/

Had to remove the Redis keycache as it Vapor 4 dropped support for it. Replaced with a custom UserCache that caches some often-used info for all users.

Vapor 4 changed how migrations work. Moved all the schema-creating migrations to a new file. Minor changes to most of the data-importing migrations.

Tests are all broken, have not started converting them to Vapor 4 as yet. Without the unit tests, this code is sure to be buggy. I’ve curl’ed a bunch of the endpoints, and can create accounts, login, and get users, twarrts, and forum categories with my client, but nothing more than that. This commit changes *something* in almost every route handler, so…there’s bugs in there somewhere.