Big change is the addition of a new way to do input validation. ValidatingJSONParser works by being a decoder that wraps a JSONDecoder, and runs validation code as decoding happens. Why do this, instead of using the built-in validation package? I wanted validation to produce a dictionary of [field : errorString] pairs, so that data could be sent back in the response packet. This lets clients associate form validation errors with the form elements that failed validation.
Anyway. This code doesn’t do the thing that is the ‘meat’ of the vapor validation package—it doesn’t autogenerate human-readable errors from tests that fail. But, this code is a lot more flexible. It allows for multi-field validation, complex validations not expressible by vapor’s validation operators, and it automatically finds and runs validations on nested structs.

Pursuant to the above, ErrorMiddleware is replaced with a new version that extends the errorResponse type to emit field errors in responses. I think I could have instead extended ErrorMiddleware with a non-default static constructor that would do the stuff I needed; there’s about 10 lines of code in ErrorMiddleware that aren’t the constructor that would have been saved.

Most request content that optionally uploads images now makes the image field an optional.