Login, CreateUser, ResetPassword, and CodeOfConduct pages.
Not done, but they aren’t looking too bad.

Made a few minor API changes: PW reset for logged in users now takes the current password and verifies it matches before changing to the new PW. And, PW recovery struct has a field for the intended new password; if everything verifies we reset the password as part of the call.