Added notification support.
API layer now has the first of its methods to return notification info in AlertsController; backed by changes to User that tracks @mentions as well as changes to post/edit/delete routes that track and update @mention counts.

UI layer adds NotificationsMiddleware, which calls the new AlertsController once per minute, saving results (and time called) in the session. So: once per minute when the user makes a GET request, we check for notification updates before servicing the request.

Leaf templates now indicate new Twarrt and Forum mentions to the user, and there are new pages for showing the mentions themselves. Navbar can also indicate the current tab.

Added a new customLeafTag: formatPostText(), which adds links to @mentions and #hashtags.

The Announcement Model is in, but the API layer CRUD methods for it are not.

ContentFilterable now has a contentTextStrings() method that collects all the text strings a ContentFilterable object should search when filtering. This way, filters can be extension methods on ContentFilterable instead of implmenting filters on each model type.

Made a new method in ForumController that handles all search types for finding ForumPosts. Removed 6 search methods that handled specific search domains.