Added identicons, changed twitarr threads
Threads in Twitarr used to be Twitter-like, in that a single tweet could have multiple direct replies, and a reply-tweet could itself be replied to. Replies therefore acted as a acyclic directed graph. Now, replies are single level. When you create a twarrt that is a reply to a twarrt that has no replies, you create a ‘reply group’ and both the reply and replied-to twarrt are added to it. Any twarrt replying to a twarrt in the reply group becomes a member of the reply group. All twarrts in a reply group have their ‘reply group’ field set to the ID of the twarrt that was first replied to.

This has 2 useful effects. First, you can always tell from a TwarrtData whether that twarrt is part of a reply group. Previously there was no way for a replied-to twarrt to indicate that it had a reply thread—this makes it easy to know when to show a “View Thread” affordance. Second, you can access the entire reply thread from the data in any twarrt that’s a member of the thread. All replies will be in the reply group, with no sub-groups.

Server-generated Identicons have been added as an optional feature clients may use. An image call was added that returns a user’s custom avatar image if one exists, else it returns an identicon for the user. Implementing this required changes to SwiftGD, so I brought the sources into the app because I am lazy and these terrible changes are never going back to the SwiftGD project.

Added user profile edit page.