Added Prometheus support, and fixed image orientation issue
Prometheus is a server metrics collector and viewer; see http://prometheus.io for information. The ‘support’ here is that Swiftarr now includes the SwiftPrometheus exporter module, and exposes a metrics endpoint for Prometheus to poll.

The image issue required reaching down into how GD Library reads and writes JPEG files. Version 2.3.2 of the GD Library ignores exif tags, meaning a read -> write of jpeg data is tag-erasing. If the original file had an exif tag specifying the image was upside-down, the output image is still upside-down but now has no tag.

A future version of GD may be fixing this problem by rotating the image data during file read. The fix I employed saves the orientation value on read and adds it back on write, but only for jpegs. This solution works for our purposes.