Context sensitive search
Tweets, Forums, and events now have search capabilities. Search bar is in the menu.

Added percentEncodeFilePathEntry() in a bunch of places. This fn protects URL parameters that get turned into path segments in the the site layer sends to the API layer.

Forums can now be favorited in the web UI.