More ForumController fns now check accessLevelToView.
Previously, categories had accessLevelToView, which allowed us to make mod-only categories and could make categories with other view requirements. But, only forum-returning methods checked this—the fn that searched for a text string in forum posts would find posts in categories the user couldn’t see and return them.

This could be fixed either by joining forums and then joining the category in db calls (so we could check category.accessLevelToView), or copying the accessLevel on creation. I chose the middle road, copying the accessLevelToView to forums but not to posts.

This means: Changing the access level of a Category could be difficult. But moving a forum from one category to another won’t be so bad, even if the categories have different access levels.

Also fixed an issue with PostSearchPageContext’s paginator.