Reconfigured several calls for performance.
Mostly, took several instances where an API calls returns an array result and had to make one or more database calls on each array element to gather data, and converted those per-element calls into single calls that operate on the entire array.

Extended Array<Model> with childCountsPerModel(), which collects counts of a related child model for each element of the array, and returns a dictionary of [ModelID : RelatedModelCount]. This method tries to use SQLKit directly, bypassing Fluent for performance, but falls back to Fluent if necessary.