Added WebSocket support.
There are 2 kinds of websockets that clients may attach to: the first is a notification socket, which receives messages for any notification that would change values in their global notification response. The second is a Fez/Seamail socket, which receives messages for each new message posted to the Fez/Seamail, and for Fezzes may also receive user join/leave messages.

Neither of these sockets have any defined messages they send to the server.

The UI also has support for the fez socket; while a Seamail or Fez thread page is open, the page will update to reflect new messages posted to the thread.