Added locustfile.
Locustfile doesn’t have all the stuff I want as yet, but it should give enough of a cross-section to be useful for load testing. This locustfile is designed to generate lots of successful requests to the server, focusing on calls users are likely to make often, although the calls aren’t weighted. Perhaps 100 endpoints get hit. The locustfile does no verification of responses.

Fixed a few minor bugs in other files.