Added pagination to Fez/Seamail list endpoints.
Previously we’d return all matching elements, which became slow for large numbers of matches. I’ll need to do the same thing for messages in a Fez/Seamail as well.

Also converted a bunch of Forum calls to use UserCache instead of User for auth.

Added some simple sanity checks at startup time. Checks that we can access the databases and that the static resource files are in the right place.