Fez and Seamail posts are now paginated.
fezHandler() in FezController now returns a paginated list of posts in a fez. UI layer paginates the fez post page, only shows the message creation UI on the last page of the list, and the websocket only attaches for the last page (since updates only happen there).

Modified the FezData DTO; added the paginator sub-struct to it.

Modified the Fez WebSocket message structure. Now, all sockets will receive structured data, and HTML sockets will also receive HTML to insert as a field in the JSON.