Modified package, attempting to make installs easier.
Also trying to make it easier to run on arm. Moved Resources and seeds directories to be inside of the App directory; this is because the Swift Package Manager really want them there in order to copy them into the App bundle directory.

Locust shows login as being kinda slow; trying to make it faster.

Several places that access the resources dir now use Bundle.module to do so.