Bundle.module turns out to be a problem.
When doing a Swift Package build, a hidden file is added to each module as it is built; that file defines Bundle.module as a module-private Bundle extension. When building from an Xcode project, it … doesn’t add this file. Also, it appears that Resources declared in the Package.swift file get copied into the app when building with SPM, but the generated Xcode project … doesn’t reference them at all.

Anyway, replaced Bundle.module with Bundle(for Settings.self), and added a new copy files phase that copies the Resources and Seeds dirs where the bundle code can find them. I *think* it works for both spm and xcodeproj builds.