Changed FezController’s postAddHandler to return FezPostData.
FezController’s postAddHandler was recently changed to just return status. It had been returning the entire FezData, which meant the entire thread got returned. That was becoming slow for large threads. However, Locust was using the postID of the returned post in a post/delete test. Now we only return the FezPostData for the new post, and Locust can grab the postID to delete it.