Performance optimization for getting lists of forum threads from the API.
forumListGetLastPosts() is used to get the last ForumPost in each forum in an array of forums. It tries to use SQLKit to make a single, complicated SQL query, but the query takes ~9ms instead of the previous ~82ms to gather results for 50 forums. It still falls back on the ORM if SQL isn’t available.