Incorporating Moderator feedback
Added 'twitarrteam' as a new access level. Members of the Twitarr dev team should be elevated to this access level. TwitarrTeam
access is between Moderator and THO; importantly it gives access to the @TwitarrTeam seamail inbox.

Added a call to promote users to TwittarTeam, callable by THO and above, and a call to promote users to THO, callable by admin.

The call to downgrade access levels requires THO access to set 'banned' or 'unverified'. Moderator level still required to set
'quarantined' or 'verified'.

Added the '?foruser=' query parameter to several Fez API calls; used by moderators and TwittarTeam members to access their
respective shared Seamail inboxes.

Added a Moderation Root page, which is just is a menu of links to more interesting pages.