Moderator Notifications
The global notification endpoint now has Moderator-specific values that are filled in for Moderator-level users. These values include the number of open reports and the number of Seamail threads to @moderator with new messages.

Also, the Moderator Guide is now available in the Moderation root page.

Also improved the active page highlighting for menu items under the Twitarr menu.