Added buttons in User Profile to go to twarrts and forum that user authored.
Does not have a button for posts authored by that user. I'm less convinced it would be useful.
As part of the above, added a new forum search method.

Also added "sticky-top" to the navbar, which gives the navbar "position: sticky" CSS. I *think* it's supported well enough by now.