Fix for issue where one user could theoretically work around blocks.
If user A blocks B, B might be able to undo the block by blocking and unblocking A. This might be difficult to do as B could no longer see A's profile page, but the API would still allow the POST to happen.

Anyway, when B unblocks A we'd been removing all the Redis blocks without checking whether A also had a block on B.

Also has a fix for a JS issue where errors weren't getting parsed because we were calling "response.text()" which 'eats' the response, preventing "response.json()" from parsing.