Fix for moderator seamail notifications
When someone posts to a seamail thread that includes @moderator, seamail notifications are added to each moderator-level user so that each mod-level user can clear the notification individually when they read the mail.

But, clearing the notification was just clearing the counts for @moderator. Fixed now.

Also fixes an issue where the userCache wasn't be updated for temp quarantined users.

Adds contentID to the NotificationSocket struct. This field gives the ID of the content entity being reported.

Also added a thing so that the Seamail thread page would keep scrolled to the bottom when reloading after posting.