Fix for schedule updating
Deleting a file when there's no file there throws an exception. This was preventing the file from getting updated.

I swear I tested this--I probably added the delete code in last, and it worked because it deleted a file that had been uploaded. The intent of the delete code is to ensure any previous schedule update files are cleaned out.