Added target to build CLI tool to modify dates in schedule.ics
The boat will be in the AST timezone on Tuesday. Sched.com doesn't appear to have the capability to list events on the same schedule in different timezones, and THO solved this by making the event times 'right' for EST, even though we'll be in AST.

This script modifies Tuesday's event times to match the actual UTC time when they happen; this means when an event time is displayed in AST it should match the ship's wall clocks.