Fix for deleting twarrts from the Twarrt Mod page.
Dunno when this broke. Perhaps when I cleaned up the JS button code?

Anyway, the Delete button works by telling the dialog about the nearest "data-postid" property, which when a user deletes their own twarrt is in the twarrt's <li>, but needs to be inserted into the Delete btn itself for the Mod page.