Async/Await refactor.
This is a largish refactor of the codebase to update to Swift 5.5's new async/await concurrency model. Most route handlers were affected, as were migrations, middleware, and extensions. However, the refactor does not involve changing the business logic of affected functions--instead, the structure changed from a promise-based series of nested callback closures to a linear async flow.

I also did a few other code cleanups that had been put off until after the 2022 sailing:
* Barrels are gone; instead there are MuteWord and AlertWord models. Bookmarks for things are now various pivot models. Blocked and muted users are now array-based fields in the User model.
* All the Redis calls are now in RedisWrapper.swift, which is designed to contain 'light' function wrappers for Redis calls that better describe the operation being performed. This makes it easier to reason about how we're using Redis.
* Did some fixups to notification handling.
* Auth is always done with UserCacheData now. Before, some route handlers authed a User, some authed a UserCacheData.
* AlertWordData and MuteWordData are now just KeywordData.

I also updated dependencies, which led to an issue with CNIOBoringSSL: CNIOBoringSSL_bn.h includes <inttypes.h>, which causes errors when put into a module. I temp-fixed it by removing the include and modifying the one line of t_x509.c that uses it; I expect the next dot release will probably fix this issue. If not, the issue can be fixed by dinking with the modulemap file.