Fixes #59
Twarrts that start a thread now have a 🧵 emoji; twarrts that are replies have a 🪡 emoji. The needle and thread emoji isn't an ideal icon for the "thread reply" concept, but it'll do until we can get custom artwork.

The Tweets -> Filters menu now has a "Hide Replies" option, which filters reply twarrts out of the stream unless you're looking at a replyGroup. This is a sticky option, saved in the sessionData, so if a user chooses to hide replies, they'll stay hidden until they choose Show All again.