Fixes #46.
Moderators can now start seamail threads as user @moderator instead of as themselves. Moderators will also be able to use live message updating (the server message socket) when viewing a message thread where @moderator is a member of the threads (and the current user isn't).

Also added an indicator to seamail thread views showing whether the socket was connected.