Navbar rework.
Twitarr in the navbar has been replaced with a Menu icon. Renamed "Seamail" to "Chat" in the navbar to save space. Added "LFG" to the navbar--previously it was only accessible in the drop down menu. Added a Search icon to the right of the menu, which only appears if the current page supports context-sensitive searching.

We'd gotten feedback that people weren't aware they could tap on Twitarr to open the menu, they had difficulty finding the Searchbar, and they felt LFG should more prominent. This addresses all of these concerns.

My worry now is that with 7 touch targets in the navbar things might be too cluttered.

Also messed with configure.swift a bit. Swiftarr doesn't actually support separate hostnames or ports for the API calls vs the UI calls; currently all calls are in the same trie router and are all served by the same HTTPServer component. I don't see a way to configure a Vapor instance to be able to run multiple HTTPServers. If we eventually split the UI and API layers into separate executables, we'll get separate addressing for free, but until then I don't want to have a bunch of Environment configuration vars that we can't actually support.

Currently the API calls are hardcoded to reside at the "/api/v3" path prefix, but this could be fixed such that the API prefix may be overridden by env vars, so I'm leaving it in.