Fixes issue #3.
The Event Forum categories now sort threads by the start time of their Event by default. The Sort dropdown has a new "Event" option, only visible in Event categories.

This commit has Model changes, and requires a db rebuild to use.
It also adds a field to the CategoryData struct in the API.