Added fullTextFilter() method to QueryBuilder.
This method acts as a drop-in replacement for query.filter() methods that do a text search on a string-valued model property, where the search string is in a local variable. For Postgres dbs, this method allows natural language search patterns to be specified in the search string. ILIKE is used for other db backends, although this method could be extended to apply other database's proprietary SQL extensions as well.

Renamed Model+Extensions to Fluent+Extensions, to better describe its contents.