Fixes #42.
Added char counters to most form textfields.

Also fixed an issue with forum creation where the forum gets created but the first post fails. Now the user is redirected to the newly created forum, and their post is is inserted into the form for editing. Previously, the user would unknowingly create multiple forums as they tried to get the post to complete without errors.