Added time zone change table.
This table as entries for every timezone change the boat goes through during the cruise. Using this table lets us determine what TimeZone we'll be in at a certain time, and improves our date calculations.

There's two kind of date calculations that use this table. One is for a Date that specifies a point in time (as Dates do). The timezone change table lets us properly format these Date objects into strings with the correct timezone. The second kind of date calculation is for Date objects that actually attempt to specify a 'floating' time, e.g. "2:00 PM tomorrow, in whatever timezone the boat is in at that time". Since Date objects are timezone-independent, they can't directly embody this concept. So, the convention is now that the database stores such dates in the ship's Port timezone--the timezone of the port we embark from. This should equal the timezone that the Sched.com schedule uses. When delivered by the API, these dates are adjusted to the Date that specifies the same clock time, in the timezone the ship was/will be in at the given time. Several API objects that have Date properties now also have a TimeZone property.

One of the benefits of this method of date handling is that Dates that are meant to indicate a 'floating' time actually work that way--if for some reason the boat changes its itinerary, we can update the TimeZoneChanges table and all the Dates in affected Event, Fez, and Announcement objects delivered by the API will be adjusted correctly.

As a result of this change:
- ScheduleMunger is no longer needed
- TimeZoneChanges has a Model, schema Migration, data import Migration, and DTO struct
- The admin API adds 2 methods: for returning the timezonechange table, and for reloading the timezone info from the seeds file
- Admin panel in the UI has a page for viewing the timezone table.