Added 'open' Seamail chats.
A new type of chat where the owner of the chat can add/remove users after creation. UI for this chat type has a clear callout telling all participants that the owner can add others at any time.

At the API level, the main change is the addition of a new FezType enum. Most of the rest of the changes to api-level files add validations and guards for the new fez type. Adding or removing users uses the same API calls as for LFGs.