Forum threads for Events now show the event start time in a separate field.
Previously the title for Event forum threads was similar to "(Sat 8:00 AM) Theme: Welcome New Cruisers!" Now the title is just the Event title, the time is separate.

The ForumListData DTO has 2 new fields that support this change: the optional eventTime, and optional timeZone. These will only be non-nil for Event Forums.