Some stuff got merged into main that shouldn't have.
The big issue is SiteFileController, where I commented out the line that adds CacheControl headers to local files. This makes it easier to work on the javascript files, but is not good for prod.

I also accidentally checked in the Logger shim that silences the missing `metadataProvider` message spam. Harmless, but no longer needed as Vapor has cleared up the spam.

Also added some stuff to the API changelist documenting the existence of the new phonecall stuff.