Got rid of ForumReaders.readCounts
ForumReaders.readCount was a way of remembering the number of posts a user has viewed in a thread, used when we calculate the number of unread posts in the thread and when we determine the default page to start on when viewing a (long) thread. We stored the value in this way because it made calculations easy. The big drawback of readCount is that it interacts poorly with changes to the visible posts in a thread. Blocking, muting, and deleted posts all change the number of visible posts in a thread, and may change the number of viewed posts.

However, now that we have childCountsPerModel() we can query the count of read posts in each thread by storing the postID of the last post the user viewed and querying the # of posts with an id <= the last viewed post. And, for routes that return an array of forum threads, it's one SQL call per request.

Also fixed a dumb bug I found in siteController where we'd divide by zero if you set limit=0 in a pagination query.

This has a new migration that modifies the ForumReaders pivot.