Calendar subscriptions (#154)
* Added support for Calendar Subscriptions

Many calendaring apps allow you to subscribe to a calendar feed, which at its simplest is an endpoint that delivers a .ics file.

This adds an endpoint that delivers a .ics file with all the events a user follows. Because calendaring apps access this endpoint in a different context than the browser, this endpoint implements the Basic Auth flow. And, to allow multiple users' calendars to be subscribed at once, the session cookies from this route are restricted to the path of the route, and that path contains the user's username.

When a calendaring app calls this new route and opens a session, we label that session "Calendaring App"; this is visible on the logout screen.

Also: Added "lastUpdateTime" to the EventData DTO; calendaring apps use this field to implement change management.

* X-WR-CALNAME now has interpolated year value

---------

Co-authored-by: Chall Fry <chall@challfry.com>