Mod page for viewing a Seamail Post fix
The page has a "View In Context" button that went to /fez/:fezID, but that doesn't work if it's a seamail. As a temp solution, added another button, and labeled each to explain where they go. In the future we'll need to modify the API-level call to identify whether the post is from a lfg or a seamail.