Updates to Swift package
Since Vapor's examples and starter projects got rid of the package design where your codebase was built as a lib and linked into a single-file app for no apparent reason,
I made a similar change to Swiftarr. I also renamed the application target to 'swiftarr' from 'App'.

This commit also updates to the latest Vapor version, which involved some minor code changes. Vapor is in the process of implementing Sendable support, which drove many of the source changes.

Also, I made a wrapper object for the methods in configure.swift; the object mostly exists to store the Logger to use while configuration takes place.