Added automated schedule updates
Uses Vapor Queues to run an hourly job to pull the JoCo schedule.ics file from Sched.com and apply schedule changes to the database.

Queues is described as being similar to Ruby Sidekiq. Although it can work as an external worker process we're not using it that way--jobs get executed in the server process without any special command line options. Queues uses Redis to coordinate jobs and track job completions and retries. The main value in doing it this way (instead of just using a Timer that fires every hour) is that the the db tracks when an hourly update is completed and won't re-do the update if the server restarts. And, since this takes a few seconds per hour, running a whole separate process (launched with the `swiftarr queues` command) is overkill.

Supporting this feature:
* The results of automatic schedule update actions are saved in a ScheduleLog, a new model added to the db
* The admin UI has a page to show the schedule update log, and a detail page to show what events were updated in any update action
* Added Queues and the QueuesRedisDriver to the package
* Added an Admin DTO struct to deliver info about schedule update actions
* Added a Settings value with the URL to query for the schedule.ics file