Added SWIFTARR_EXTERNAL_URL env var.
This environment variable is the externally-routable URL to the server--what a user would enter into a browser address bar to get to Swiftarr.

It's used by the server to build URLs pointing to itself. Defaults to "http://" plus the first hostname in the "SWIFTARR_CANONICAL_HOSTNAMES" env var if available, or "http://localhost:8081" as a backup.