Forum Search improvements, Search UI refactor
`/api/v3/forum/search` gains a bunch of new filter options, including filter by category, which makes context-sensitive search on category pages viable.

The web UI's search bar works differently now. Instead of having multiple search bar configurations in the trunk with a cascading #if to choose one, pages with an in-menu search bar put the template for the search bar in their page under a separate #export, and the trunk #imports it in the right place. This will make it easier to make custom search UI on specific pages.