Added Photostream feature; reworked main menu
First, the pulldown menu in the upper-left is gone, all of its menu items are now on the Home page. Hopefully this will simplify navigation.

Second, the home screen now has a Photostream cell that shows a rotating carousel of photos tagged with the place onboard where the photo was taken (and the timestamp for the photo).

Photos should only be added to the Photostream by native apps, as we want uploaded photos to be freshly taken by the camera and not pulled from a user's photo roll. Native apps should also use OCR to blur out any text identified in the photo.