Removing the SwiftNIO concurrency executor
These lines in entrypoint.swift set SwiftNIO as the global concurrency executor, which roughly means "the thing that chooses what thread async continuations run on". I haven't looked into it, but most likely SwiftNIO tries to keep continuations running on the thread/queue where they started, instead of running them on a global operations queue. The default executor is almost certainly designed for apps and tries to move anything it can off the 'main' thread, and any code that isn't MainActor isolated can run on background threads.

This has the effect of removing the precondition code in SwiftNIO that is crashing during Redis shutdown.