Added a new class of registration codes for Discord users
These codes are only for preproduction instances of Twitarr and will not function on boat. These registration codes are a separate pool marked for this use case, and there's admin UI to 'allocate' a code by tying it to a particular user on the JoCo Cruise discord. The use case is a user on Discord who wants to check out Twitarr before the cruise (often to familiarize themselves with the moderation features). Now any TwitarrTeam member can allocate a registration code for that Discord user and DM them a reg code they can use to create an account. And, unlike previous years, we'll have at least some idea who that Twitarr user is on the pre-production server.

API changes: A few admin-level structs were changed: RegistrationCodeUserData has new fields used for RegistrationCodes marked for Discord users. RegistrationCodeStatsData has new fields showing totals for Discord users.

This commit has 2 new migrations which must be run for these new features to work.