Fixes #224.
The @admin, @THO, @TwitarrTeam, and @moderator accounts now have distinctive styling in bylines and @mentions. Styling is based on the username and not access level, as the UserHeader is the only info we have on the user at the time we're styling their name.

This means that moderator users won't have their bylines highlighted. This is both because it's more involved of a change (we'd probably have to add a userLevel field to UserHeader), and also because mod users probably want to post as 'normal' users most of the time, and can post as @moderator (and get the special styling) when they need to.