Fixes #257.
ImageController now checks that a file actually exists at the filepath where it should be, and 404s if there's no file there. Previously we'd pass the path into `streamFile()`, which would return a 500 level error as it assumes it's always getting valid files to stream.

Also fixed up the favicon handling.