Making CodeOfConduct into a Markdown file
For some time, the Code Of Conduct has been delivered to us as a JSON file with ambiguous parsing rules. Unless there's some spec I'm unaware of, JSON by itself isn't a great choice for a document markup language.

I'm hoping that we can remove the CodeOfConduct.json file and exclusively use the .md file this year, but the first step is creating a viable alternative.