Added simple server rollup feature.
The admin panel now has a 'Show Table Counts' link that shows a bunch of row counts for various database tables. The counts are computed on request, and there's no facility for saving counts or showing hourly/daily diffs. Still, this makes it easier to see a quick overview of what's being created on the server and track feature usage.