Day Planner and Personal Events
Day Planner is now available from the Home screen, showing a calendar-style view of the day's schedule. This view combines followed Events, LFGs that the user has created or joined, and Private Events they've created or been invited to. Also has a list view for private events.

Private Events are very similar to Open Seamail chats, except they have a place and time associated with them. The UI for showing them is pretty much the Seamail UI with a few extra fields. The 'Private' in private events specifically means these events aren't advertised to non-members they way LFGs are.

Reminder events solely for the creator are now called Personal Events, are just another type of LFG internally, and are mostly differentiated by not allowing users to join and not allowing posts.

In the near future I want to add stuff to FezData or FezType at the UI level to support capability queries (e.g. "canAddUsers" or "shouldShowStartTime"). This would replace a bunch of tests against the FezType where we check for multiple different types, when what we really want is all the types that support a capability.

Also added a Former tab to the LFG panel. This shows the user LFGs, Seamail chats, and Personal Events the user was previously a member of, but no longer. The primary purpose of this feature is to allow users to report on chats they've been kicked from, although the user can use the view to re-join LFGs they were previously in.

The `fezListItem` leaf template (shows a cell-sized preview of a LFG) can now be used to show LFGs, Seamail Chats, and Personal Events, meaning you can take a mixed list of all three types and show them all appropriately with a simple loop.

Trunk now has a `#import("subheader")` allowing for pages that want a sticky subheader under the navbar.

LFGs now store their time in the port timezone, and ingest times assuming the given Date is actually a floating time in the current timezone (not the tz the boat will be in when the LFG actually happens). Similar to Events, we then convert times to the tz the boat will be in at the start time when delivering FezDatas.