Improved notifications for Private Events
This commit adds fields to UserNotificationData for Private Event chat types and also for counts of chat types the user has been added to. There is now a separate field for the # of Private Event chats with new messages. There are also fields for the number of chats of each type the user has been added to. This notification only shows cases where the chat owner adds the user, not for chats you create yourself or LFGs you join yourself. If a user has been added to a chat and hasn't yet viewed it, that chat will be included in the "Added To" total and not the "New Messages" total, even if the chat also has new messages.

This commit also adds new Notification Socket message types for users who get added to chats.

Also expands on the "Notifications" pill in the Home view, now with more notification types shown.