Personal Events UI (#327)
* Making the timezone change table a no-auth route.

It was an Admin-only route because only Admin needed it (to view and update the table), but now Kraken needs the info, and there's no actual reason to protect get GET route.

* The main event-getting route now considers 'days' to start at midnight and end at 3 AM the next day.

This means an event starting at 1:00 AM Monday may appear on Sunday's schedule page and also on Monday's schedule page. That's okay.

The various options for applying time filters to the Schedule queries also handle time zones a bit better.

* PersonalEvents are now a new LFG type.

This gives Personal Events their own chat so participants may discuss their event.

There's actually 2 new LFG types: `personalEvent` is only viewable by its creator; other users may not be added and there's no chat. This is most similar to adding an appointment to your phone's calendar. `privateEvent` may have other users added, and works in a very similar fashion to an Open Seamail chat. UI clients should show the Start Time, End Time, and Location to participants, probably near the chat view.

The existing PersonalEventController routes are still there; they now mostly call through to FezController to do their work.

There's also some changes to how LFGs handle time zones.

This also has a fix for the bug where a user couldn't report a Seamail chat because they got removed from it. There is now a 'Former' tab in LFG, which list LFGs, Open Seamails, and Private Events the user was formerly a member of. The user cannot see the contents of the chats (they're not members anymore), but hey can report them.

* Working on Dayplanner UI.

CSS tables are the best. I'm learning so many things about the box model.

* Day Planner and Personal Events

Day Planner is now available from the Home screen, showing a calendar-style view of the day's schedule. This view combines followed Events, LFGs that the user has created or joined, and Private Events they've created or been invited to. Also has a list view for private events.

Private Events are very similar to Open Seamail chats, except they have a place and time associated with them. The UI for showing them is pretty much the Seamail UI with a few extra fields. The 'Private' in private events specifically means these events aren't advertised to non-members they way LFGs are.

Reminder events solely for the creator are now called Personal Events, are just another type of LFG internally, and are mostly differentiated by not allowing users to join and not allowing posts.

In the near future I want to add stuff to FezData or FezType at the UI level to support capability queries (e.g. "canAddUsers" or "shouldShowStartTime"). This would replace a bunch of tests against the FezType where we check for multiple different types, when what we really want is all the types that support a capability.

Also added a Former tab to the LFG panel. This shows the user LFGs, Seamail chats, and Personal Events the user was previously a member of, but no longer. The primary purpose of this feature is to allow users to report on chats they've been kicked from, although the user can use the view to re-join LFGs they were previously in.

The `fezListItem` leaf template (shows a cell-sized preview of a LFG) can now be used to show LFGs, Seamail Chats, and Personal Events, meaning you can take a mixed list of all three types and show them all appropriately with a simple loop.

Trunk now has a `#import("subheader")` allowing for pages that want a sticky subheader under the navbar.

LFGs now store their time in the port timezone, and ingest times assuming the given Date is actually a floating time in the current timezone (not the tz the boat will be in when the LFG actually happens). Similar to Events, we then convert times to the tz the boat will be in at the start time when delivering FezDatas.

* Cleaned up JS, removed test code

* Added LFGLabel Leaf tag, added PEs back to Seamail

I've been going back and forth on this, but I think Private Events (the ones with chats) need to be displayed in the Seamail tab.

I also added a new Leaf tag to improve our display of the various kinds of LFGs.

* Missed a file with last checkin.

* Improved notifications for Private Events

This commit adds fields to UserNotificationData for Private Event chat types and also for counts of chat types the user has been added to. There is now a separate field for the # of Private Event chats with new messages. There are also fields for the number of chats of each type the user has been added to. This notification only shows cases where the chat owner adds the user, not for chats you create yourself or LFGs you join yourself. If a user has been added to a chat and hasn't yet viewed it, that chat will be included in the "Added To" total and not the "New Messages" total, even if the chat also has new messages.

This commit also adds new Notification Socket message types for users who get added to chats.

Also expands on the "Notifications" pill in the Home view, now with more notification types shown.

---------

Co-authored-by: Chall Fry <chall@challfry.com>